Implement cryptorgraphy

+++ src/main/kotlin/ir/roudi/directory/DirectoryRouting.kt
@@ -18,7 +18,7 @@ fun Application.configureDirectoryRouting() {
             val params = call.receiveParameters()
             val name = params.getOrFail("name")
             val port = params.getOrFail("port")
-            Node.create(name, port).let {
+            Node.create(name, port.toInt()).let {
                 DirectoryDataSource.saveNode(it)
             }
             call.respondText("Node saved successfully!", status = HttpStatusCode.Created)
